rename todo.js file
